+++ PrivateFrameworks/iWorkImport.framework/TNPageLayout.h
@@ -40,7 +40,7 @@
 - (void)addChild:(id)arg1;
 - (struct CGSize { double x1; double x2; })adjustedInsetsForTarget:(id)arg1;
 - (bool)alwaysStartsNewTarget;
-- (unsigned int)autosizeFlagsForTextLayout:(id)arg1;
+- (unsigned long long)autosizeFlagsForTextLayout:(id)arg1;
 - (struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })autosizedFrameForTextLayout:(id)arg1 textSize:(struct CGSize { double x1; double x2; })arg2;
 - (struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })bodyRect;
 - (unsigned long long)columnCount;
