+++ PrivateFrameworks/iWorkImport.framework/TSWPStorageAttributeRangeProvider.h
@@ -7,7 +7,7 @@
 }
 
 - (void)dealloc;
-- (id)initWithStorage:(id)arg1 kind:(int)arg2;
+- (id)initWithStorage:(id)arg1 kind:(unsigned long long)arg2;
 - (void)nextRange;
 
 @end
