+++ PrivateFrameworks/iWorkImport.framework/TSWPStorageStyler.h
@@ -3,20 +3,20 @@
  */
 
 @interface TSWPStorageStyler : NSObject {
-    int  _attributeArrayKind;
+    unsigned long long  _attributeArrayKind;
     TSWPStorage * _storage;
 }
 
-@property (nonatomic) int attributeArrayKind;
+@property (nonatomic) unsigned long long attributeArrayKind;
 @property (nonatomic, retain) TSWPStorage *storage;
 
 - (void).cxx_destruct;
 - (void)applyStyle:(id)arg1 range:(struct _NSRange { unsigned long long x1; unsigned long long x2; })arg2 undoTransaction:(struct TSWPStorageTransaction { }*)arg3;
-- (int)attributeArrayKind;
-- (id)initWithStorage:(id)arg1 attributeArrayKind:(int)arg2;
+- (unsigned long long)attributeArrayKind;
+- (id)initWithStorage:(id)arg1 attributeArrayKind:(unsigned long long)arg2;
 - (void)performStylingInRange:(struct _NSRange { unsigned long long x1; unsigned long long x2; })arg1 undoTransaction:(struct TSWPStorageTransaction { }*)arg2;
 - (id)replacementStyleForStyle:(id)arg1 range:(struct _NSRange { unsigned long long x1; unsigned long long x2; })arg2;
-- (void)setAttributeArrayKind:(int)arg1;
+- (void)setAttributeArrayKind:(unsigned long long)arg1;
 - (void)setStorage:(id)arg1;
 - (id)storage;
